requirements.txt
Flask==2.0.1
flask_sqlalchemy==2.5.1
flask_login==0.5.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
python-dotenv==0.19.0
Werkzeug==2.0.3
SQLAlchemy==1.3.23
